Added save load game support with local device independence

+++ Assets/Scripts/Core/Datastructure/GameData.cs
@@ -8,7 +8,7 @@ namespace Test.FactoryRun.Core
     {
         public static Func<int, bool> UpdateGem;
         
-        public int Gems;
+        public double Gems;
 
         public GameData() 
         {
